Adding my community contribution
I couldn't figure out how best to do brightness, so I opted for differentiating walls by color instead. I also colored the rays according to which wall each ray intersects. Mostly I just wanted to fix the rendering math. In the original coding challenge, you had a slider for the field of view of the camera. Here, I have a slider for the focal length of the camera. I also draw the camera in the top down view on the left. Oh, and use WASD to move, because the arrow keys kept scrolling the window and that got annoying.